Rename MruEvictionStrategy to MruEvictionManager
