+++ src/evictionmanagers/MruEvictionManager.ts
@@ -22,4 +22,8 @@ export class MruEvictionManager<K extends PK> implements IEvictionManager<K> {
       this.order.splice(index, 1);
     }
   }
+
+  clear(): void {
+    this.order = [];
+  }
 }
